Add xz as a dependency where genreflex is required
The LCG 96 release installed ROOT with an external xz rather than the
builtin, so the genreflex binary needs to find it.
Normally this could be added to the LD_LIBRARY_PATH in the root
configuration to setup its dependent environment, however as we
install ROOT as an external package (taken from cvmfs) we cannot
access to the xz prefix, hence we need to manually add it on the
rest of packages.

As a side note, xz is taken also from cvmfs but even if we put it
on the RPATH it is not found by genreflex.